Oracle analysis for different lengths, and it visibilized bugg of n>2M made the code crash. added asserts to fails of cuBlas calls, will fix by batching cuBlasCalls

+++ code/cuda/.ipynb_checkpoints/GPUDataManager-checkpoint.h
@@ -22,7 +22,7 @@ class GPUDataManager {  //Class to handle the data allocation and movement for t
         void freeData(DeviceData *pdevData, DeviceData *d_pdevData);
     private:
         void createOnesVec(DeviceData *pdevData);
-        unsigned int onesVecLen;
+        unsigned long onesVecLen;
     
 };
